api/requirements.txt
fastapi
uvicorn[standard]
pillow
face_recognition
numpy
qdrant-client
redis
aioredis
python-multipart
dlib
pika
aio-pika